refactor AzureClientManager code and update pytests
