+++ backend/manage-indexing-jobs.py
@@ -17,9 +17,9 @@ from kubernetes import (
     config,
 )
 
-from src.api.azure_clients import AzureClientManager
 from src.logger.load_logger import load_pipeline_logger
 from src.typing.pipeline import PipelineJobState
+from src.utils.azure_clients import AzureClientManager
 from src.utils.common import sanitize_name
 from src.utils.pipeline import PipelineJob
