+++ backend/src/api/prompt_tuning.py
@@ -13,8 +13,8 @@ from fastapi import (
 )
 from graphrag.config.create_graphrag_config import create_graphrag_config
 
-from src.api.azure_clients import AzureClientManager
 from src.logger.load_logger import load_pipeline_logger
+from src.utils.azure_clients import AzureClientManager
 from src.utils.common import sanitize_name
 
 prompt_tuning_route = APIRouter(prefix="/index/config", tags=["Index Configuration"])
