+++ backend/src/logger/__init__.py
@@ -6,7 +6,6 @@ from src.logger.application_insights_workflow_callbacks import (
 )
 from src.logger.console_workflow_callbacks import ConsoleWorkflowCallbacks
 from src.logger.load_logger import load_pipeline_logger
-from src.logger.logger_singleton import LoggerSingleton
 from src.logger.pipeline_job_updater import PipelineJobUpdater
 from src.logger.typing import (
     Logger,
@@ -18,7 +17,6 @@ __all__ = [
     "Logger",
     "ApplicationInsightsWorkflowCallbacks",
     "ConsoleWorkflowCallbacks",
-    "LoggerSingleton",
     "PipelineAppInsightsReportingConfig",
     "PipelineJobUpdater",
     "PipelineReportingConfigTypes",
