+++ backend/src/utils/azure_clients.py
